+++ pages/_campus/dashboard.vue
@@ -53,7 +53,6 @@
       <bulma-tile vertical>
         <has-phone-tile :data="stats[REQUESTABLE.hasPhone]" />
         <statuses-tile :statuses="stats[REQUESTABLE.statuses]" />
-        <ratings-tile :ratings="stats[REQUESTABLE.ratings]" />
       </bulma-tile>
       <bulma-tile vertical>
         <pois-tile
@@ -89,7 +88,6 @@ import hasPhoneTile from '~/components/dashboard/has-phone.vue';
 import overTimeTile from '~/components/dashboard/over-time.vue';
 import poisTile from '~/components/dashboard/pois.vue';
 import statusesTile from '~/components/dashboard/statuses.vue';
-import ratingsTile from '~/components/dashboard/ratings.vue';
 
 const REQUESTABLE = {
   total: 'total',
@@ -101,7 +99,6 @@ const REQUESTABLE = {
   drivers: 'drivers',
   hasPhone: 'has-phone',
   period: 'period',
-  ratings: 'ratings',
 };
 
 export default {
@@ -114,7 +111,6 @@ export default {
     hasPhoneTile,
     statusesTile,
     overTimeTile,
-    ratingsTile,
   },
   props: {
     campus: {
@@ -139,7 +135,7 @@ export default {
           `${REQUESTABLE.poisDeparture}(id,departure(location(coordinates),label),total)`,
           `${REQUESTABLE.poisArrival}(id,arrival(location(coordinates),label),total)`,
           `${REQUESTABLE.drivers}(id,driver(name,firstname,lastname),total)`,
-          REQUESTABLE.period, REQUESTABLE.ratings,
+          REQUESTABLE.period,
         ].join(','),
       },
       start,