+++ CHANGELOG.md
@@ -1,10 +1,8 @@
 # Changelog
 ## Current
 ### Features
-* Added ratings list by campus page
-* Added rating details page
-* Added ride details page
 * Added loading animation and success message while creating or editing CRUD
+* Added alert message if browser is not Firefox
 ### Fix
 * Add missing create button in Phone index
 * Add missing create button in POI index
@@ -15,6 +13,8 @@
 * Fixed driver name display in history table
 * Fix SSR rendering of pages where calendar is present
 * Fixed rights to import csv and create new data for phones and pois
+* Fixed csv export mask to fit import template
+* Added format-coordinates.js to generate number compatible format
 ## Version 1.4.5
 ### Features
 * Add driver connection status in planificator